Add more css to experiment with positioning
I've positioned everything so that it's in the middle
of the webpage. I've also tried to play around with padding
and margins, with little success.

There's also an issue with the bullet point, as it's on the
left side, whereas its text is in the centre. However, I
did notice that when I changed the margin or padding values,
it started to move.